requirements.txt
yfinance==0.2.37
pandas==2.1.4
numpy==1.26.4
scipy==1.12.0
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.4.0
torch==2.2.0
torchvision==0.17.0
pyyaml==6.0.1
ta==0.11.0
torchsummary==1.5.1
tqdm==4.66.1
joblib==1.3.2
